requirements.txt
pandas
requests
gtfs-realtime-bindings
protobuf
psycopg2-binary